[#42560801] Continuação do novo layout / Novo mover (janelinha, árvore de pastas destino, salvar, verificação de alguma pasta selecionada, atualização dos dados da 'janelinha') / Locales

+++ app/controllers/lesson_files_controller.rb
@@ -1,5 +1,6 @@
 class LessonFilesController < ApplicationController
 
+  layout "define_token"
   require 'fileutils' # utilizado na remoção de diretórios, pois o "Dir.rmdir" não remove diretórios que não estejam vazis
 
 
@@ -8,11 +9,13 @@ class LessonFilesController < ApplicationController
 
     if @lesson and @lesson.type_lesson == Lesson_Type_File
       # file_default = lesson.address
-      file    = File.join(Lesson::FILES_PATH, params[:lesson_id])
-      @files  = directory_hash(file, @lesson.name).to_json if File.exist?(file)
+      file     = File.join(Lesson::FILES_PATH, params[:lesson_id])
+      Dir.mkdir(file) unless File.exist?(file)
+      @files   = directory_hash(file, @lesson.name).to_json
+      @folders = directory_hash(file, @lesson.name, false).to_json
     end
 
-    render :layout => "define_token"
+    # render :layout => "define_token"
   end
 
   def new
@@ -25,25 +28,48 @@ class LessonFilesController < ApplicationController
         error = true
       end
 
+      @lesson = Lesson.where(id: params[:lesson_id]).first
+      file     = File.join(Lesson::FILES_PATH, params[:lesson_id])
+      @files   = directory_hash(file, @lesson.name).to_json
+      @folders = directory_hash(file, @lesson.name, false).to_json
+
       respond_to do |format|
-        format.html{ render :nothing => true, :status => (error ? 500 : 200) }
+        format.html{ render (error ? {:nothing => true} : :index), :status => (error ? 500 : 200) }
       end
     end
   end
 
   def edit
+
     begin
-      path       = File.join("#{Rails.root}", "media", "lessons", params[:lesson_id], params[:path])
-      path_split = path.split(File::SEPARATOR)               #[media, lessons, lesson_id, path_parte1, path_parte2]
-      path_split.delete(path_split.last)                     #[media, lessons, lesson_id, path_parte1]
-      new_path   = File.join(path_split, params[:node_name]) #/media/lessons/lesson_id/path_parte1/node_name]
-      FileUtils.mv path, new_path # renomeia
-    rescue Exception => error
+      if params[:type] == "rename" # renomear
+
+        path       = File.join("#{Rails.root}", "media", "lessons", params[:lesson_id], params[:path])
+        path_split = path.split(File::SEPARATOR)               #[media, lessons, lesson_id, path_parte1, path_parte2]
+        path_split.delete(path_split.last)                     #[media, lessons, lesson_id, path_parte1]
+        new_path   = File.join(path_split, params[:node_name]) #/media/lessons/lesson_id/path_parte1/node_name]
+        FileUtils.mv path, new_path # renomeia
+
+      elsif params[:type] == "move" # mover
+        params[:paths_to_move].each do |node_path|
+          path     = File.join("#{Rails.root}", "media", "lessons", params[:lesson_id], node_path)
+          new_path = File.join("#{Rails.root}", "media", "lessons", params[:lesson_id], params[:path_to_move_to])
+          FileUtils.mv path, new_path if File.exist?(path) # move
+        end
+
+      end
+
+    rescue
       error = true
     end
 
+    @lesson  = Lesson.where(id: params[:lesson_id]).first
+    file     = File.join(Lesson::FILES_PATH, params[:lesson_id])
+    @files   = directory_hash(file, @lesson.name).to_json
+    @folders = directory_hash(file, @lesson.name, false).to_json
+
     respond_to do |format|
-      format.html{ render :nothing => true, :status => (error ? 500 : 200) }
+      format.html{ render (error ? {:nothing => true} : :index), :status => (error ? 500 : 200) }
     end
   end
 
@@ -57,14 +83,19 @@ class LessonFilesController < ApplicationController
       error = true
     end
 
+    @lesson = Lesson.where(id: params[:lesson_id]).first
+    file     = File.join(Lesson::FILES_PATH, params[:lesson_id])
+    @files   = directory_hash(file, @lesson.name).to_json
+    @folders = directory_hash(file, @lesson.name, false).to_json
+
     respond_to do |format|
-      format.html{ render :nothing => true, :status => (error ? 500 : 200) }
+      format.html{ render (error ? {:nothing => true} : :index), :status => (error ? 500 : 200) }
     end
   end
 
   private
 
-    def directory_hash(path, name=nil)
+    def directory_hash(path, name=nil, get_children=true)
       data = {title: (name || path)}
       data[:children] = children = []
       data[:isFolder] = true
@@ -72,8 +103,8 @@ class LessonFilesController < ApplicationController
         next if ['.', '..'].include?(entry)
         full_path = File.join(path, entry)
         if File.directory?(full_path)
-          children << directory_hash(full_path, entry)
-        else
+          children << directory_hash(full_path, entry, get_children)
+        elsif get_children
           children << entry
         end
       end
